didnt realize the file system would have the new changes as well
It came to my attention when testing on s99 that the file system
that is explored in my getOtherHostHeaderValuesFromCluster function
would have the file that is being checked within the deployment
paramter. This change skips that orders file from being checked,
because of course it would have the FORWARD_HOST_HEADERS variable.