forward host header duplicate check (#148)
* put this in a dockerfile because shellcheck npm is linux version

* working tests - not complete

* running it locally on a mac

* setup loglevel for running tests locally

I for one would love to see the log.info that I have setup and this aims to aid in that. We can now actually set the loglevel via environment, where before it was being ignored.

* passing no dupe test

* separate test for single vs multiple values

* test for unique value

* test occasionally fails due to the order changing

this sorts the results alphabetically so tests consistently pass

* dont output this...

* skip this check on jobs clusters

* code review with Shawn

* not needed param

* better name

* didnt realize the file system would have the new changes as well

It came to my attention when testing on s99 that the file system
that is explored in my getOtherHostHeaderValuesFromCluster function
would have the file that is being checked within the deployment
paramter. This change skips that orders file from being checked,
because of course it would have the FORWARD_HOST_HEADERS variable.

* summarize the noDuplicateForwardHostHeaders check